requirements.txt
flask
python-dotenv
groq
pypdf
docx2txt
faiss-cpu
sentence-transformers
numpy
